+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailViewModel.kt
@@ -20,12 +20,7 @@ import android.net.Uri
 import androidx.annotation.IdRes
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.airbnb.mvrx.Async
-import com.airbnb.mvrx.Fail
-import com.airbnb.mvrx.FragmentViewModelContext
-import com.airbnb.mvrx.MvRxViewModelFactory
-import com.airbnb.mvrx.Success
-import com.airbnb.mvrx.ViewModelContext
+import com.airbnb.mvrx.*
 import com.jakewharton.rxrelay2.BehaviorRelay
 import com.jakewharton.rxrelay2.PublishRelay
 import com.squareup.inject.assisted.Assisted
@@ -68,6 +63,7 @@ import im.vector.riotx.core.utils.subscribeLogError
 import im.vector.riotx.features.command.CommandParser
 import im.vector.riotx.features.command.ParsedCommand
 import im.vector.riotx.features.home.room.detail.timeline.helper.TimelineDisplayableEvents
+import im.vector.riotx.features.home.room.typing.TypingHelper
 import im.vector.riotx.features.settings.VectorPreferences
 import io.reactivex.Observable
 import io.reactivex.functions.BiFunction
@@ -84,6 +80,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                                                       userPreferencesProvider: UserPreferencesProvider,
                                                       private val vectorPreferences: VectorPreferences,
                                                       private val stringProvider: StringProvider,
+                                                      private val typingHelper: TypingHelper,
                                                       private val session: Session
 ) : VectorViewModel<RoomDetailViewState, RoomDetailAction>(initialState), Timeline.Listener {
 
@@ -93,16 +90,16 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
     private val visibleEventsObservable = BehaviorRelay.create<RoomDetailAction.TimelineEventTurnsVisible>()
     private val timelineSettings = if (userPreferencesProvider.shouldShowHiddenEvents()) {
         TimelineSettings(30,
-                         filterEdits = false,
-                         filterTypes = true,
-                         allowedTypes = TimelineDisplayableEvents.DEBUG_DISPLAYABLE_TYPES,
-                         buildReadReceipts = userPreferencesProvider.shouldShowReadReceipts())
+                filterEdits = false,
+                filterTypes = true,
+                allowedTypes = TimelineDisplayableEvents.DEBUG_DISPLAYABLE_TYPES,
+                buildReadReceipts = userPreferencesProvider.shouldShowReadReceipts())
     } else {
         TimelineSettings(30,
-                         filterEdits = true,
-                         filterTypes = true,
-                         allowedTypes = TimelineDisplayableEvents.DISPLAYABLE_TYPES,
-                         buildReadReceipts = userPreferencesProvider.shouldShowReadReceipts())
+                filterEdits = true,
+                filterTypes = true,
+                allowedTypes = TimelineDisplayableEvents.DISPLAYABLE_TYPES,
+                buildReadReceipts = userPreferencesProvider.shouldShowReadReceipts())
     }
 
     private var timelineEvents = PublishRelay.create<List<TimelineEvent>>()
@@ -160,6 +157,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
 
     override fun handle(action: RoomDetailAction) {
         when (action) {
+            is RoomDetailAction.UserIsTyping                     -> handleUserIsTyping(action)
             is RoomDetailAction.SaveDraft                        -> handleSaveDraft(action)
             is RoomDetailAction.SendMessage                      -> handleSendMessage(action)
             is RoomDetailAction.SendMedia                        -> handleSendMedia(action)
@@ -237,32 +235,41 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                         copy(
                                 // Create a sendMode from a draft and retrieve the TimelineEvent
                                 sendMode = when (draft) {
-                                               is UserDraft.REGULAR -> SendMode.REGULAR(draft.text)
-                                               is UserDraft.QUOTE   -> {
-                                                   room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
-                                                       SendMode.QUOTE(timelineEvent, draft.text)
-                                                   }
-                                               }
-                                               is UserDraft.REPLY   -> {
-                                                   room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
-                                                       SendMode.REPLY(timelineEvent, draft.text)
-                                                   }
-                                               }
-                                               is UserDraft.EDIT    -> {
-                                                   room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
-                                                       SendMode.EDIT(timelineEvent, draft.text)
-                                                   }
-                                               }
-                                           } ?: SendMode.REGULAR("")
+                                    is UserDraft.REGULAR -> SendMode.REGULAR(draft.text)
+                                    is UserDraft.QUOTE   -> {
+                                        room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
+                                            SendMode.QUOTE(timelineEvent, draft.text)
+                                        }
+                                    }
+                                    is UserDraft.REPLY   -> {
+                                        room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
+                                            SendMode.REPLY(timelineEvent, draft.text)
+                                        }
+                                    }
+                                    is UserDraft.EDIT    -> {
+                                        room.getTimeLineEvent(draft.linkedEventId)?.let { timelineEvent ->
+                                            SendMode.EDIT(timelineEvent, draft.text)
+                                        }
+                                    }
+                                } ?: SendMode.REGULAR("")
                         )
                     }
                 }
                 .disposeOnClear()
     }
 
+    private fun handleUserIsTyping(action: RoomDetailAction.UserIsTyping) {
+        if (vectorPreferences.sendTypingNotifs()) {
+            if (action.isTyping) {
+                room.userIsTyping()
+            } else {
+                room.userStopsTyping()
+            }
+        }
+    }
+
     private fun handleTombstoneEvent(action: RoomDetailAction.HandleTombstoneEvent) {
-        val tombstoneContent = action.event.getClearContent().toModel<RoomTombstoneContent>()
-                               ?: return
+        val tombstoneContent = action.event.getClearContent().toModel<RoomTombstoneContent>() ?: return
 
         val roomId = tombstoneContent.replacementRoom ?: ""
         val isRoomJoined = session.getRoom(roomId)?.roomSummary()?.membership == Membership.JOIN
@@ -400,7 +407,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                 is SendMode.EDIT    -> {
                     // is original event a reply?
                     val inReplyTo = state.sendMode.timelineEvent.root.getClearContent().toModel<MessageContent>()?.relatesTo?.inReplyTo?.eventId
-                                    ?: state.sendMode.timelineEvent.root.content.toModel<EncryptedEventContent>()?.relatesTo?.inReplyTo?.eventId
+                            ?: state.sendMode.timelineEvent.root.content.toModel<EncryptedEventContent>()?.relatesTo?.inReplyTo?.eventId
                     if (inReplyTo != null) {
                         // TODO check if same content?
                         room.getTimeLineEvent(inReplyTo)?.let {
@@ -409,13 +416,13 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                     } else {
                         val messageContent: MessageContent? =
                                 state.sendMode.timelineEvent.annotations?.editSummary?.aggregatedContent.toModel()
-                                ?: state.sendMode.timelineEvent.root.getClearContent().toModel()
+                                        ?: state.sendMode.timelineEvent.root.getClearContent().toModel()
                         val existingBody = messageContent?.body ?: ""
                         if (existingBody != action.text) {
                             room.editTextMessage(state.sendMode.timelineEvent.root.eventId ?: "",
-                                                 messageContent?.type ?: MessageType.MSGTYPE_TEXT,
-                                                 action.text,
-                                                 action.autoMarkdown)
+                                    messageContent?.type ?: MessageType.MSGTYPE_TEXT,
+                                    action.text,
+                                    action.autoMarkdown)
                         } else {
                             Timber.w("Same message content, do not send edition")
                         }
@@ -426,7 +433,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                 is SendMode.QUOTE   -> {
                     val messageContent: MessageContent? =
                             state.sendMode.timelineEvent.annotations?.editSummary?.aggregatedContent.toModel()
-                            ?: state.sendMode.timelineEvent.root.getClearContent().toModel()
+                                    ?: state.sendMode.timelineEvent.root.getClearContent().toModel()
                     val textMsg = messageContent?.body
 
                     val finalText = legacyRiotQuoteText(textMsg, action.text.toString())
@@ -542,7 +549,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
             when (val tooBigFile = attachments.find { it.size > maxUploadFileSize }) {
                 null -> room.sendMedias(attachments)
                 else -> _fileTooBigEvent.postValue(LiveEvent(FileTooBigError(tooBigFile.name
-                                                                             ?: tooBigFile.path, tooBigFile.size, maxUploadFileSize)))
+                        ?: tooBigFile.path, tooBigFile.size, maxUploadFileSize)))
             }
         }
     }
@@ -731,8 +738,7 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
                 .buffer(1, TimeUnit.SECONDS)
                 .filter { it.isNotEmpty() }
                 .subscribeBy(onNext = { actions ->
-                    val bufferedMostRecentDisplayedEvent = actions.maxBy { it.event.displayIndex }?.event
-                                                           ?: return@subscribeBy
+                    val bufferedMostRecentDisplayedEvent = actions.maxBy { it.event.displayIndex }?.event ?: return@subscribeBy
                     val globalMostRecentDisplayedEvent = mostRecentDisplayedEvent
                     if (trackUnreadMessages.get()) {
                         if (globalMostRecentDisplayedEvent == null) {
@@ -795,7 +801,14 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
         room.rx().liveRoomSummary()
                 .unwrap()
                 .execute { async ->
-                    copy(asyncRoomSummary = async)
+                    val typingRoomMembers =
+                            typingHelper.toTypingRoomMembers(async.invoke()?.typingRoomMemberIds, room)
+
+                    copy(
+                            asyncRoomSummary = async,
+                            typingRoomMembers = typingRoomMembers,
+                            typingMessage = typingHelper.toTypingMessage(typingRoomMembers)
+                    )
                 }
     }
 
@@ -882,6 +895,9 @@ class RoomDetailViewModel @AssistedInject constructor(@Assisted initialState: Ro
     override fun onCleared() {
         timeline.dispose()
         timeline.removeAllListeners()
+        if (vectorPreferences.sendTypingNotifs()) {
+            room.userStopsTyping()
+        }
         super.onCleared()
     }
 }